Add AccountMenuIcon component to AppBarTop and LandingSiteNav
